sparesults::solution

Struct QuerySolution

pub struct QuerySolution { /* private fields */ }

Tuple associating variables and terms that are the result of a SPARQL query.

It is the equivalent of a row in SQL.

use sparesults::QuerySolution;
use oxrdf::{Variable, Literal};

let solution = QuerySolution::from((vec![Variable::new("foo")?, Variable::new("bar")?], vec![Some(Literal::from(1).into()), None]));
assert_eq!(solution.get("foo"), Some(&Literal::from(1).into())); // Get the value of the variable ?foo if it exists (here yes).
assert_eq!(solution.get(1), None); // Get the value of the second column if it exists (here no).

Implementations

impl QuerySolution

pub fn get(&self, index: impl VariableSolutionIndex) -> Option<&Term>

Returns a value for a given position in the tuple (usize) or a given variable name (&str, Variable or VariableRef).

use sparesults::QuerySolution;
use oxrdf::{Variable, Literal};

let solution = QuerySolution::from((vec![Variable::new("foo")?, Variable::new("bar")?], vec![Some(Literal::from(1).into()), None]));
assert_eq!(solution.get("foo"), Some(&Literal::from(1).into())); // Get the value of the variable ?foo if it exists (here yes).
assert_eq!(solution.get(1), None); // Get the value of the second column if it exists (here no).

pub fn len(&self) -> usize

The number of variables which could be bound.

It is also the number of columns in the solutions table.

use oxrdf::{Literal, Variable};
use sparesults::QuerySolution;

let solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![Some(Literal::from(1).into()), None],
));
assert_eq!(solution.len(), 2);

pub fn is_empty(&self) -> bool

Is there any variable bound in the table?

use oxrdf::{Literal, Variable};
use sparesults::QuerySolution;

let solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![Some(Literal::from(1).into()), None],
));
assert!(!solution.is_empty());

let empty_solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![None, None],
));
assert!(empty_solution.is_empty());

pub fn iter(&self) -> impl Iterator<Item = (&Variable, &Term)>

Returns an iterator over bound variables.

use oxrdf::{Literal, Variable};
use sparesults::QuerySolution;

let solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![Some(Literal::from(1).into()), None],
));
assert_eq!(
    solution.iter().collect::<Vec<_>>(),
    vec![(&Variable::new("foo")?, &Literal::from(1).into())]
);

pub fn values(&self) -> &[Option<Term>]

Returns the ordered slice of variable values.

use oxrdf::{Literal, Variable};
use sparesults::QuerySolution;

let solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![Some(Literal::from(1).into()), None],
));
assert_eq!(solution.values(), &[Some(Literal::from(1).into()), None]);

pub fn variables(&self) -> &[Variable]

Returns the ordered slice of the solution variables, bound or not.

use oxrdf::{Literal, Variable};
use sparesults::QuerySolution;

let solution = QuerySolution::from((
    vec![Variable::new("foo")?, Variable::new("bar")?],
    vec![Some(Literal::from(1).into()), None],
));
assert_eq!(
    solution.variables(),
    &[Variable::new("foo")?, Variable::new("bar")?]
);